:recycle: `ResultInfoBox` 컴포넌트 리팩토링 #70 - 게임 결과 관련 type, api 파일 이동
