+++ src/constants/player.ts
@@ -1,0 +1,1 @@
+export const PLAYER_COLORS = ['pink', 'green', 'yellow', 'blue'] as const;
