+++ src/pages/result/api.ts
@@ -1,0 +1,5 @@
+import { client } from '@/features/api';
+
+export const getGameResult = async () => {
+  return client.get<GameResult>('/ingame').then(({ data }) => data);
+};
