+++ src/pages/result/constants.ts
@@ -47,5 +47,3 @@ export const MODAL_REWARDBOX =
 export const TIER_UPGRADE_MESSAGE = 'Origin에 한 발자국 더 나아갔습니다!';
 
 export const REWARD_MESSAGE = '잃어버린 색을 찾았습니다!';
-
-export const COLOR_STYLES = ['pink', 'yellow', 'green', 'blue'] as const;
