+++ src/pages/result/index.ts
@@ -1,0 +1,2 @@
+export { default } from './page';
+export { getGameResult } from './api';
