+++ src/pages/result/page.tsx
@@ -3,15 +3,11 @@ import { useLoaderData, useNavigate } from 'react-router-dom';
 
 import * as constants from './constants';
 import * as styles from './index.css';
-
-import type { GameResult } from '@/types/result';
+import ResultInfoBox from './ResultInfoBox';
+import { RewardsModal } from './RewardsModal';
 
 import ColoredButton from '@/components/button/ColoredButton';
-
-import useModal from '@/hooks/useModal';
-
-import ResultInfoBox from '@/pages/result/ResultInfoBox';
-import { RewardsModal } from '@/pages/result/RewardsModal/index';
+import Modal, { useModal } from '@/components/modal';
 
 import { getOut } from '@/services/lobby';
 
@@ -22,7 +18,7 @@ const Result = () => {
   const navigate = useNavigate();
   const { roomUuid, players, reward } = useLoaderData() as GameResult;
   const myCode = useUserCode();
-  const { Modal, openModal, closeModal } = useModal();
+  const { modalRef, openModal, closeModal } = useModal();
   const hasTier = !!reward.tier && reward.tier.upgrade;
   const noReward = !hasTier && reward.skins.length === 0;
 
@@ -52,19 +48,12 @@ const Result = () => {
       <div className={styles.gameResultWrapper}>
         <div className={styles.result}>
           <div className={styles.resultTitle}>{constants.RESULTTEXT}</div>
-
           {players.map((item) => (
-            <div key={`${item.nickname}${item.rank}`}>
-              <ResultInfoBox
-                rank={item.rank}
-                imgSrc={item.skin}
-                label={item.label}
-                nickname={item.nickname}
-                colorStyle={constants.COLOR_STYLES[item.index]}
-                gameScore={item.score}
-                myResult={item.userCode === myCode}
-              />
-            </div>
+            <ResultInfoBox
+              key={item.userCode}
+              player={item}
+              myResult={item.userCode === myCode}
+            />
           ))}
         </div>
 
@@ -85,7 +74,7 @@ const Result = () => {
       </div>
 
       {!noReward && (
-        <Modal>
+        <Modal ref={modalRef}>
           <RewardsModal
             tier={reward.tier}
             skin={reward.skins}
