+++ src/pages/result/types.d.ts
@@ -1,3 +1,19 @@
-import * as resultConstants from './constants';
+type PlayerRecord = {
+  index: number;
+  score: number;
+} & Pick<User, 'nickname' | 'label' | 'skin' | 'rank' | 'userCode'>;
 
-type colorStyles = (typeof resultConstants.COLOR_STYLES)[number];
+interface TierReward {
+  oldtier: tierRange;
+  newtier: tierRange;
+  upgrade: boolean;
+}
+
+interface GameResult {
+  roomUuid?: string;
+  players: PlayerRecord[];
+  reward: {
+    tier?: TierReward;
+    skins: string[];
+  };
+}
